+++ src/functions/config.ts
@@ -1,17 +1,18 @@
-import { APP_NAME, I_LIKE_JS } from "../constants.ts";
+import { APP_NAME, DEFAULT_SETTINGS, I_LIKE_JS } from "../constants.ts";
 import type { CONFIG_FILES } from "../types.ts";
 import { CheckForPath, JoinPaths } from "./filesystem.ts";
 import { LogStuff } from "./io.ts";
+import { stringify as stringifyYaml } from '@std/yaml';
 
 /**
  * Returns file paths for all config files the app uses.
  *
  * @export
- * @param {("BASE" | "MOTHERFKRS" | "LOGS" | "UPDATES")} path What path you want.
+ * @param {("BASE" | "MOTHERFKRS" | "LOGS" | "UPDATES" | "SETTINGS")} path What path you want.
  * @returns {string} The path as a string.
  */
 export async function GetAppPath(
-    path: "BASE" | "MOTHERFKRS" | "LOGS" | "UPDATES",
+    path: "BASE" | "MOTHERFKRS" | "LOGS" | "UPDATES" | "SETTINGS",
 ): Promise<string> {
     const appDataPath = Deno.env.get("APPDATA");
     if (!appDataPath) {
@@ -24,7 +25,8 @@ export async function GetAppPath(
     const BASE_DIR = await JoinPaths(appDataPath, APP_NAME.CLI);
     const PROJECTS = await JoinPaths(BASE_DIR, `${APP_NAME.CLI}-${I_LIKE_JS.MFS.toLowerCase().replace("*", "o").replace("*", "u")}.txt`);
     const LOGS = await JoinPaths(BASE_DIR, `${APP_NAME.CLI}-logs.log`);
-    const UPDATES = await JoinPaths(BASE_DIR, `${APP_NAME.CLI}-updates.json`);
+    const UPDATES = await JoinPaths(BASE_DIR, `${APP_NAME.CLI}-updates.yaml`);
+    const SETTINGS = await JoinPaths(BASE_DIR, `${APP_NAME.CLI}-settings.yaml`);
 
     switch (path) {
         case "BASE":
@@ -35,6 +37,8 @@ export async function GetAppPath(
             return LOGS;
         case "UPDATES":
             return UPDATES;
+        case "SETTINGS":
+            return SETTINGS;
         default:
             throw new Error("Invalid path requested");
     }
@@ -64,10 +68,18 @@ export async function FreshSetup(): Promise<CONFIG_FILES> {
             await Deno.writeTextFile(logsPath, "");
         }
 
+        const updatesPath = await GetAppPath("UPDATES");
+
+        const settingsPath = await GetAppPath("SETTINGS");
+        if (!(await CheckForPath(settingsPath))) {
+            await Deno.writeTextFile(settingsPath, stringifyYaml(DEFAULT_SETTINGS))
+        }
+
         return {
             projects: projectPath,
             logs: logsPath,
-            updates: await GetAppPath("UPDATES"),
+            updates: updatesPath,
+            settings: settingsPath,
         };
     } catch (e) {
         await LogStuff(`Some ${I_LIKE_JS.MFN} error happened trying to setup config files: ${e}`, "error");