+++ src/main.ts
@@ -13,6 +13,15 @@ import { ColorString, LogStuff, ParseFlag } from "./functions/io.ts";
 import { FreshSetup } from "./functions/config.ts";
 import GenericErrorHandler from "./utils/error.ts";
 
+async function init(update: boolean) {
+    const configFiles = await FreshSetup(); // Temporarily hold the result
+    const ALL_CONFIG_FILES: CONFIG_FILES = configFiles; // Assign only after it's resolved
+    await TheUpdater({ CF: ALL_CONFIG_FILES, force: update, silent: !update });
+    return ALL_CONFIG_FILES;
+}
+
+const ALL_CONFIG_FILES: CONFIG_FILES = await init(false);
+
 const [inputCommand] = Deno.args;
 
 if (!inputCommand) {
@@ -27,15 +36,6 @@ const flags = Deno.args.map((arg) => {
 
 const isVerbose = flags.includes("--verbose");
 const wantsToUpdate = flags.includes("--update");
-
-async function init(update: boolean) {
-    const configFiles = await FreshSetup(); // Temporarily hold the result
-    const ALL_CONFIG_FILES: CONFIG_FILES = configFiles; // Assign only after it's resolved
-    await TheUpdater({ CF: ALL_CONFIG_FILES, force: update, silent: !update });
-    return ALL_CONFIG_FILES;
-}
-
-const ALL_CONFIG_FILES = await init(false);
 
 if (ParseFlag("help", true).some((flag) => flags.includes(flag))) {
     await init(false);